drm/nouveau : fix compiler warning about incompatible types
JIRA: https://issues.redhat.com/browse/RHEL-1350

Upstream Status: RHEL-only

+++ drivers/gpu/drm/nouveau/nouveau_drv.h
@@ -187,7 +187,7 @@ u_free(void *addr)
 }
 
 static inline void *
-u_memcpya(uint64_t user, unsigned int nmemb, unsigned int size)
+u_memcpya(uint64_t user, size_t nmemb, size_t size)
 {
 	void __user *userptr = u64_to_user_ptr(user);
 	size_t bytes;
